chore: sync hotfix v0.32.1 to staging

+++ src/db/schema.ts
@@ -47,7 +47,7 @@ export const runsCosts = pgTable(
     costName: text("cost_name").notNull(),
     costSource: text("cost_source").notNull(),
     quantity: numeric("quantity", { precision: 20, scale: 6 }).notNull(),
-    unitCostInUsdCents: numeric("unit_cost_in_usd_cents", { precision: 12, scale: 10 }).notNull(),
+    unitCostInUsdCents: numeric("unit_cost_in_usd_cents", { precision: 16, scale: 10 }).notNull(),
     totalCostInUsdCents: numeric("total_cost_in_usd_cents", { precision: 16, scale: 10 }).notNull(),
     status: text("status").notNull().default("actual"),
     idempotencyKey: text("idempotency_key"),
